Corrected a small numerical error when calculating the number of 8052 instructions to execute in a unit of time. Fixed the problem of rouding errors causing zero instructions to execute, which results in an infinite loop.

+++ src/qemu/qqq_rpc.h
@@ -98,7 +98,7 @@ class uCsim_Machine:
 		int read_pipe[2];
 		int write_pipe[2];
 		static const double mega_hz;
-		static const int instrs_per_usec;
+		static const double instrs_per_usec;
 		pthread_mutex_t mtx;
 		char run_buf[1000];
 		char write_buf[1000];
